% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{anglesArc}
\alias{anglesArc}
\title{Compute the angular limits of an arc}
\usage{
anglesArc(vector, theta)
}
\arguments{
\item{vector}{Unit direction vector from an arc centre.}

\item{theta}{Half-angle of the arc in radians.}
}
\value{
A numeric vector of start and end angles in radians.
}
\description{
Compute the angular limits of an arc
}
\examples{
anglesArc(c(1, 0), pi / 4)
}
