% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{dummycoor}
\alias{dummycoor}
\title{Construct an exterior circumcentre for a boundary edge}
\usage{
dummycoor(tri.obj, firstPoint, secondPoint, centre, away)
}
\arguments{
\item{tri.obj}{A \code{triSht} object returned by \code{\link[interp:tri.mesh]{interp::tri.mesh()}}.}

\item{firstPoint, secondPoint}{Numeric coordinate vectors defining an edge.}

\item{centre}{Circumcentre of the adjacent Delaunay triangle.}

\item{away}{Distance used to place the exterior point.}
}
\value{
Numeric coordinate vector for the exterior circumcentre.
}
\description{
Reproduces \code{alphahull::dummycoor}, testing the exterior side of a boundary
edge against the convex hull of a \code{triSht} triangulation.
}
\examples{
tri.obj <- interp::tri.mesh(c(0, 1, 0), c(0, 0, 1))
dummycoor(
  tri.obj,
  firstPoint = c(0, 0), secondPoint = c(1, 0),
  centre = c(0.5, 0.5), away = 1
)
}
