% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{inter}
\alias{inter}
\title{Calculate the intersection geometry of two circles}
\usage{
inter(c11, c12, r1, c21, c22, r2)
}
\arguments{
\item{c11, c12}{Centre coordinates of the first circle.}

\item{r1}{Radius of the first circle.}

\item{c21, c22}{Centre coordinates of the second circle.}

\item{r2}{Radius of the second circle.}
}
\value{
A list with the number of intersections, two direction vectors, and
the corresponding half-angles.
}
\description{
Calculate the intersection geometry of two circles
}
\examples{
inter(0, 0, 1, 1, 0, 1)
}
