% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{lengthahull}
\alias{lengthahull}
\title{Calculate alpha-hull arc length}
\usage{
lengthahull(arcs)
}
\arguments{
\item{arcs}{Matrix or data frame containing \code{theta} and \code{r} columns.}
}
\value{
Total arc length.
}
\description{
Calculate alpha-hull arc length
}
\examples{
arcs <- data.frame(theta = c(pi / 4, pi / 2), r = c(2, 1))
lengthahull(arcs)
}
