% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pointInPolygon}
\alias{pointInPolygon}
\title{Test whether a point lies in a polygon}
\usage{
pointInPolygon(x, y, polygonX, polygonY)
}
\arguments{
\item{x, y}{Coordinates of the query point.}

\item{polygonX, polygonY}{Coordinates of polygon vertices.}
}
\value{
\code{TRUE} when the point is inside the polygon, otherwise \code{FALSE}.
}
\description{
Test whether a point lies in a polygon
}
\examples{
pointInPolygon(0.5, 0.5, c(0, 1, 1, 0), c(0, 0, 1, 1))
}
