% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{rotation}
\alias{rotation}
\title{Rotate a two-dimensional vector}
\usage{
rotation(vector, theta)
}
\arguments{
\item{vector}{Numeric vector of length two.}

\item{theta}{Rotation angle in radians.}
}
\value{
A rotated numeric vector of length two.
}
\description{
Rotate a two-dimensional vector
}
\examples{
rotation(c(1, 0), pi / 2)
}
