% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delvor.R
\name{delvor}
\alias{delvor}
\title{Build the Delaunay mesh used by alpha-shape algorithms}
\usage{
delvor(x, y = NULL)
}
\arguments{
\item{x, y}{Coordinates accepted by \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}.}
}
\value{
An object of class \code{delvor} containing the Delaunay mesh.
}
\description{
Build the Delaunay mesh used by alpha-shape algorithms
}
\examples{
coordinates <- rbind(c(0, 0), c(1, 0), c(1, 1), c(0, 1), c(0.5, 0.5))
triangulation <- delvor(coordinates)
head(triangulation$mesh)
}
