% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicAlphaHull-package.R
\docType{package}
\name{DynamicAlphaHull-package}
\alias{DynamicAlphaHull}
\alias{DynamicAlphaHull-package}
\title{DynamicAlphaHull: adaptive alpha-hull range construction}
\description{
\code{DynamicAlphaHull} constructs range polygons from longitude-latitude
occurrence records. Its primary interface is \code{getDynamicAlphaHull()}, which
returns a \code{terra::SpatVector}.
}
\author{
\strong{Maintainer}: Yuxuan Wang \email{wyx1743@gmail.com} (\href{https://orcid.org/0009-0003-7817-748X}{ORCID})

Authors:
\itemize{
  \item Yuxuan Wang \email{wyx1743@gmail.com} (\href{https://orcid.org/0009-0003-7817-748X}{ORCID})
}

}
\keyword{internal}
